Change suffixes in Hydrogen components
