+++ packages/hydrogen/src/components/CartLineQuantityAdjustButton/tests/CartLineQuantityAdjustButton.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import {CartLineProvider} from '../../CartLineProvider';
 import {CartLineQuantity} from '../../CartLineQuantity';
-import {CartLineQuantityAdjustButton} from '../CartLineQuantityAdjustButton';
+import {CartLineQuantityAdjustButton} from '../CartLineQuantityAdjustButton.shared';
 import {CART_LINE} from '../../CartLineProvider/tests/fixtures';
 import {useCart} from '../../CartProvider';
 import {CART_WITH_LINES_FLATTENED} from '../../CartProvider/tests/fixtures';
