+++ packages/hydrogen/src/components/ExternalVideo/index.ts
@@ -1,1 +1,1 @@
-export {ExternalVideo} from './ExternalVideo';
+export {ExternalVideo} from './ExternalVideo.shared';
