+++ packages/hydrogen/src/components/ExternalVideo/tests/ExternalVideo.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import {mount} from '@shopify/react-testing';
-import {ExternalVideo} from '../ExternalVideo';
+import {ExternalVideo} from '../ExternalVideo.shared';
 import {getExternalVideoData} from '../../../utilities/tests/media';
 
 describe('<ExternalVideo />', () => {
