+++ packages/hydrogen/src/components/Image/Image.shared.tsx
