+++ packages/hydrogen/src/components/Image/index.ts
@@ -1,1 +1,1 @@
-export {Image} from './Image';
+export {Image} from './Image.shared';
