+++ packages/hydrogen/src/components/MediaFile/MediaFile.shared.tsx
