+++ packages/hydrogen/src/components/MediaFile/index.ts
@@ -1,1 +1,1 @@
-export {MediaFile, MediaFileProps} from './MediaFile';
+export {MediaFile, MediaFileProps} from './MediaFile.shared';
