+++ packages/hydrogen/src/components/Video/index.ts
@@ -1,1 +1,1 @@
-export {Video} from './Video';
+export {Video} from './Video.shared';
