+++ packages/hydrogen/src/components/Video/tests/Video.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import {mount} from '@shopify/react-testing';
-import {Video} from '../Video';
+import {Video} from '../Video.shared';
 
 const VIDEO_PROPS = {
   id: 'video',
