+++ packages/hydrogen/src/foundation/ServerRequestProvider/index.ts
@@ -1,1 +1,1 @@
-export * from './ServerRequestProvider';
+export * from './ServerRequestProvider.server';
