+++ packages/hydrogen/src/foundation/ServerStateProvider/ServerStateProvider.shared.tsx
