+++ packages/hydrogen/src/foundation/ServerStateProvider/index.ts
@@ -1,2 +1,8 @@
-export {ServerStateProvider, ServerStateContext} from './ServerStateProvider';
-export type {ServerState, ServerStateContextValue} from './ServerStateProvider';
+export {
+  ServerStateProvider,
+  ServerStateContext,
+} from './ServerStateProvider.shared';
+export type {
+  ServerState,
+  ServerStateContextValue,
+} from './ServerStateProvider.shared';
